Calculate total sum of all sub-objects splitting them by risk types

+++ src/main/java/com/kputimcevs/insurance/premium/entities/Policy.java
@@ -15,7 +15,7 @@ public class Policy {
 
     public static class SubObject {
         public String subObjectName;
-        public double sumInsured;
+        public float sumInsured;
         public RiskType riskType;
     }
 }
